+++ spec/dummy/app/models/person.rb
@@ -1,4 +1,14 @@
-require "binary_serializer"
+module BinarySerializer
+  def self.encode(raw)
+    return raw if raw.blank?
+    raw.unpack("B*")[0]
+  end
+
+  def self.decode(raw)
+    return raw if raw.blank?
+    [raw].pack("B*")
+  end
+end
 
 class Person < ActiveRecord::Base
   include Vault::EncryptedModel