+++ spec/dummy/config/initializers/vault.rb
@@ -1,8 +1,8 @@
-require "vault/rails"
+require "vault/active_record"
 
 require_relative "../../../support/vault_server"
 
-Vault::Rails.configure do |vault|
+Vault::ActiveRecord.configure do |vault|
   vault.application = "dummy"
   vault.address     = RSpec::VaultServer.address
   vault.token       = RSpec::VaultServer.token